Fix a typo in the recipes for leather backpack
Tiny change to make it require string and chest without metadata. Before it wanted a String with metadata of 1 for crafting, and same for the chest, with this change it requires a metadata of both to 0 in the recipe (vanilla item and craftable). This way the (new) leather backpack is now craftable